+++ test/integration/startPage.js
@@ -24,6 +24,10 @@ describe('Start page', () => {
     expect($('.button__start').prop('href')).to.equal(`${constants.siteRoot}${routes.check.path}`);
   });
 
+  it('has a link to the GP finder', () => {
+    expect($('.nhsuk-page-content').find('a').prop('href')).to.equal('https://www.nhs.uk/Service-Search/GP/LocationSearch/4');
+  });
+
   it('has an urgent help call out', () => {
     expect($('.samaritans__call').text()).to.equal('116 123');
     expect($('.samaritans__email').text()).to.equal('jo@samaritans.org');