+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Do not display online referral message when no option available
 - Hyperlink search again link on no results page
 - Add extra info to search page about use of GP
+- Update content to latest tested
 
 0.5.0 / 2018-08-02
 ==================