+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Set background to blue
 - Replace references to `NHS.UK` in favour of `NHS`
 - Update header search placeholder text
+- Remove aria-labelledby on the aria-hidden NHS logo
 
 0.5.0 / 2018-08-02
 ==================