+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 - Hyperlink search again link on no results page
 - Add extra info to search page about use of GP
 - Update content to latest tested
+- Give each page a unique title
+- Set background to blue
+- Replace references to `NHS.UK` in favour of `NHS`
+- Update header search placeholder text
+- Remove aria-labelledby on the aria-hidden NHS logo
 
 0.5.0 / 2018-08-02
 ==================